server/requirements.txt
fastapi
uvicorn
qdrant-client
google-generativeai